requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
pymongo==4.7.2
motor==3.4.0
chromadb==0.5.0
google-generativeai==0.5.4
beautifulsoup4==4.12.3
requests==2.31.0
python-dotenv==1.0.1
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
httpx==0.27.0
lxml==5.2.2
pydantic==2.7.1